Step2: Code along
Very interesting trick with duplication of a custom element and a native
element that is invisible. Let the native element handle the events, and
the custom element handle the style.

Also interesting with the adjacent sibling combination. How to have a
hover on a parent impact the child!
+ for direct sibling (only when nothing between)
~ for anywhere down the sibling path (no matter what's in between)

IconWrapper technique is okay.

Glad I watched instead of struggling too much on my own.